refactor<servos> cleaner code for GPIO class
Moved it to another file in BYODR utils folder. This folder is new in the pi's docker container.

It doesn't make sense to move the whole directory where the docker compose file is. Camera files are moved also in servos docker image. They aren't used there.
